backend/requirements.txt
websockets>=11.0.3
numpy>=1.24.3
pyaudio>=0.2.13
faster-whisper>=0.9.0
python-dotenv>=1.0.0
fastapi>=0.95.2
uvicorn>=0.22.0
openai>=0.27.8
pydantic>=1.10.7